feat: cutover from EJS to React SPA
- Remove all legacy EJS routers and views (login, register, verify, users, groups, sites, external-domains, jobs, settings, apikeys, reset-password, nodes, containers)
- Extract nginx-conf and dnsmasq template endpoints to routers/templates.js
- Mount client/dist as static + SPA fallback for non-API, non-template routes
- Drop connect-flash, method-override deps (no longer needed)
- Keep views/nginx-conf.ejs and views/dnsmasq/* (server-side configuration templates)
