+++ create-a-container/package.json
@@ -18,7 +18,6 @@
   "dependencies": {
     "argon2": "^0.44.0",
     "axios": "^1.15.2",
-    "connect-flash": "^0.1.1",
     "cookie-parser": "^1.4.7",
     "csrf-csrf": "^4.0.3",
     "dotenv": "^17.2.3",
@@ -27,7 +26,6 @@
     "express-rate-limit": "^8.5.1",
     "express-session": "^1.19.0",
     "express-session-sequelize": "^2.3.0",
-    "method-override": "^3.0.0",
     "morgan": "^1.10.1",
     "nodemailer": "^8.0.5",
     "pg": "^8.16.3",
