--- create-a-container/routers/sites.js
@@ -1,253 +1,0 @@
-const express = require('express');
-const { Site, Node, Container, Service, HTTPService, TransportService, ExternalDomain } = require('../models');
-const { requireAuth, requireAdmin, requireLocalhostOrAdmin, setCurrentSite, isApiRequest } = require('../middlewares');
-
-const router = express.Router();
-
-// Shared query for dnsmasq endpoints — site with nodes and running containers
-async function loadDnsmasqSite(siteId) {
-  return Site.findByPk(siteId, {
-    include: [{
-      model: Node,
-      as: 'nodes',
-      include: [{
-        model: Container,
-        as: 'containers',
-        where: { status: 'running' },
-        required: false,
-        attributes: ['macAddress', 'ipv4Address', 'hostname']
-      }]
-    }]
-  });
-}
-
-// GET /sites/:siteId/dnsmasq/:file - Dnsmasq configuration files
-const DNSMASQ_TEMPLATES = ['conf', 'dhcp-hosts', 'hosts', 'dhcp-opts', 'servers'];
-router.get('/:siteId/dnsmasq/:file', requireLocalhostOrAdmin, async (req, res) => {
-  const { file } = req.params;
-  if (!DNSMASQ_TEMPLATES.includes(file)) return res.status(404).send('Not found');
-  const site = await loadDnsmasqSite(parseInt(req.params.siteId, 10));
-  if (!site) return res.status(404).send('Site not found');
-  res.set('Content-Type', 'text/plain');
-  return res.render(`dnsmasq/${file}`, { site });
-});
-
-// GET /sites/:siteId/nginx - Endpoint for nginx configuration
-router.get('/:siteId/nginx', requireLocalhostOrAdmin, async (req, res) => {
-  const siteId = parseInt(req.params.siteId, 10);
-  
-  // fetch services for the specific site (only from running containers)
-  const site = await Site.findByPk(siteId, {
-    include: [{
-      model: Node,
-      as: 'nodes',
-      include: [{
-        model: Container,
-        as: 'containers',
-        where: { status: 'running' },
-        required: false,
-        include: [{
-          model: Service,
-          as: 'services',
-          include: [
-            {
-              model: HTTPService,
-              as: 'httpService',
-              include: [{
-                model: ExternalDomain,
-                as: 'externalDomain'
-              }]
-            },
-            {
-              model: TransportService,
-              as: 'transportService'
-            }
-          ]
-        }]
-      }]
-    }, {
-      model: ExternalDomain,
-      as: 'externalDomains'
-    }]
-  });
-
-  // Flatten services from site→nodes→containers→services
-  const allServices = [];
-  site?.nodes?.forEach(node => {
-    node?.containers?.forEach(container => {
-      container?.services?.forEach(service => {
-        // Add container reference for template compatibility
-        service.Container = container;
-        allServices.push(service);
-      });
-    });
-  });
-  
-  // Filter by type
-  const httpServices = allServices.filter(s => s.type === 'http');
-  const streamServices = allServices.filter(s => s.type === 'transport');
-
-  // Collect domains used by HTTP services on this site + domains owned by this site
-  const usedDomainIds = new Set();
-  httpServices.forEach(s => {
-    if (s.httpService?.externalDomain?.id) usedDomainIds.add(s.httpService.externalDomain.id);
-  });
-  (site?.externalDomains || []).forEach(d => usedDomainIds.add(d.id));
-  const externalDomains = await ExternalDomain.findAll({ where: { id: [...usedDomainIds] } });
-  
-  res.set('Content-Type', 'text/plain');
-  return res.render('nginx-conf', { httpServices, streamServices, externalDomains });
-});
-
-// Apply auth to all routes below this point
-router.use(requireAuth);
-
-// store the current site for routes with :siteId
-router.use('/:siteId', setCurrentSite);
-
-// Mount sub-routers
-const nodesRouter = require('./nodes');
-const containersRouter = require('./containers');
-router.use('/:siteId/nodes', nodesRouter);
-router.use('/:siteId/containers', containersRouter);
-
-// GET /sites - List all sites (available to all authenticated users)
-router.get('/', async (req, res) => {
-  const sites = await Site.findAll({
-    include: [{
-      model: Node,
-      as: 'nodes',
-      attributes: ['id', 'name']
-    }],
-    order: [['id', 'ASC']]
-  });
-
-  const rows = sites.map(s => ({
-    id: s.id,
-    name: s.name,
-    internalDomain: s.internalDomain,
-    dhcpRange: s.dhcpRange,
-    gateway: s.gateway,
-    nodeCount: s.nodes ? s.nodes.length : 0
-  }));
-
-  if (isApiRequest(req)) {
-    return res.json({ sites: rows });
-  }
-
-  return res.render('sites/index', {
-    rows,
-    req
-  });
-});
-
-// GET /sites/new - Display form for creating a new site (admin only)
-router.get('/new', requireAdmin, async (req, res) => {
-  res.render('sites/form', {
-    site: null,
-    isEdit: false,
-    req
-  });
-});
-
-// GET /sites/:id/edit - Display form for editing an existing site (admin only)
-router.get('/:id/edit', requireAdmin, async (req, res) => {
-  const site = await Site.findByPk(req.params.id);
-  
-  if (!site) {
-    await req.flash('error', 'Site not found');
-    return res.redirect('/sites');
-  }
-
-  res.render('sites/form', {
-    site,
-    isEdit: true,
-    req
-  });
-});
-
-// POST /sites - Create a new site (admin only)
-router.post('/', requireAdmin, async (req, res) => {
-  try {
-    const { name, internalDomain, dhcpRange, subnetMask, gateway, dnsForwarders, externalIp } = req.body;
-    
-    await Site.create({
-      name,
-      internalDomain,
-      dhcpRange,
-      subnetMask,
-      gateway,
-      dnsForwarders,
-      externalIp: externalIp || null
-    });
-
-    await req.flash('success', `Site ${name} created successfully`);
-    return res.redirect('/sites');
-  } catch (error) {
-    console.error('Error creating site:', error);
-    await req.flash('error', 'Failed to create site: ' + error.message);
-    return res.redirect('/sites/new');
-  }
-});
-
-// PUT /sites/:id - Update an existing site (admin only)
-router.put('/:id', requireAdmin, async (req, res) => {
-  try {
-    const site = await Site.findByPk(req.params.id);
-    
-    if (!site) {
-      await req.flash('error', 'Site not found');
-      return res.redirect('/sites');
-    }
-
-    const { name, internalDomain, dhcpRange, subnetMask, gateway, dnsForwarders, externalIp } = req.body;
-    
-    await site.update({
-      name,
-      internalDomain,
-      dhcpRange,
-      subnetMask,
-      gateway,
-      dnsForwarders,
-      externalIp: externalIp || null
-    });
-
-    await req.flash('success', `Site ${name} updated successfully`);
-    return res.redirect('/sites');
-  } catch (error) {
-    console.error('Error updating site:', error);
-    await req.flash('error', 'Failed to update site: ' + error.message);
-    return res.redirect(`/sites/${req.params.id}/edit`);
-  }
-});
-
-// DELETE /sites/:id - Delete a site (admin only)
-router.delete('/:id', requireAdmin, async (req, res) => {
-  try {
-    const site = await Site.findByPk(req.params.id, {
-      include: [{ model: Node, as: 'nodes' }]
-    });
-    
-    if (!site) {
-      await req.flash('error', 'Site not found');
-      return res.redirect('/sites');
-    }
-
-    if (site.nodes && site.nodes.length > 0) {
-      await req.flash('error', 'Cannot delete site with associated nodes');
-      return res.redirect('/sites');
-    }
-
-    const siteName = site.name;
-    await site.destroy();
-
-    await req.flash('success', `Site ${siteName} deleted successfully`);
-    return res.redirect('/sites');
-  } catch (error) {
-    console.error('Error deleting site:', error);
-    await req.flash('error', 'Failed to delete site: ' + error.message);
-    return res.redirect('/sites');
-  }
-});
-
-module.exports = router;
