--- create-a-container/views/apikeys/show.ejs
@@ -1,82 +1,0 @@
-<%- include('../layouts/header', { 
-  title: 'API Key Details - MIE',
-  breadcrumbs: [
-    { label: 'API Keys', url: '/apikeys' },
-    { label: 'Details', url: '#' }
-  ],
-  req
-}) %>
-
-<div class="row justify-content-center">
-  <div class="col-md-8">
-    <div class="card shadow-sm">
-      <div class="card-body">
-        <div class="d-flex justify-content-between align-items-center mb-4">
-          <h4 class="card-title mb-0">API Key Details</h4>
-          <form method="POST" action="/apikeys/<%= apiKey.id %>" style="display: inline;" onsubmit="return confirm('Are you sure you want to delete this API key? This action cannot be undone.');">
-            <input type="hidden" name="_method" value="DELETE">
-            <button type="submit" class="btn btn-danger btn-sm">
-              <i class="bi bi-trash"></i> Delete Key
-            </button>
-          </form>
-        </div>
-
-        <div class="alert alert-info" role="alert">
-          <i class="bi bi-info-circle"></i>
-          For security, the full API key is not stored and cannot be displayed again. 
-          Only the key prefix is shown for identification.
-        </div>
-
-        <dl class="row">
-          <dt class="col-sm-4">Key Prefix:</dt>
-          <dd class="col-sm-8">
-            <code class="fs-5" style="word-break: break-all;"><%= apiKey.keyPrefix %>********</code>
-          </dd>
-
-          <dt class="col-sm-4">Description:</dt>
-          <dd class="col-sm-8">
-            <% if (apiKey.description) { %>
-              <%= apiKey.description %>
-            <% } else { %>
-              <span class="text-muted fst-italic">No description provided</span>
-            <% } %>
-          </dd>
-
-          <dt class="col-sm-4">Key ID:</dt>
-          <dd class="col-sm-8">
-            <code style="word-break: break-all;"><%= apiKey.id %></code>
-          </dd>
-
-          <dt class="col-sm-4">Created:</dt>
-          <dd class="col-sm-8">
-            <%= new Date(apiKey.createdAt).toLocaleString() %>
-          </dd>
-
-          <dt class="col-sm-4">Last Updated:</dt>
-          <dd class="col-sm-8">
-            <%= new Date(apiKey.updatedAt).toLocaleString() %>
-          </dd>
-
-          <dt class="col-sm-4">Last Used:</dt>
-          <dd class="col-sm-8 mb-0">
-            <% if (apiKey.lastUsedAt) { %>
-              <%= new Date(apiKey.lastUsedAt).toLocaleString() %>
-              <span class="badge bg-success">Active</span>
-            <% } else { %>
-              <span class="text-muted fst-italic">Never used</span>
-              <span class="badge bg-secondary">Inactive</span>
-            <% } %>
-          </dd>
-        </dl>
-
-        <div class="d-flex justify-content-between mt-4">
-          <a href="/apikeys" class="btn btn-secondary">
-            <i class="bi bi-arrow-left"></i> Back to List
-          </a>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
-<%- include('../layouts/footer') %>
