--- create-a-container/views/containers/index.ejs
@@ -1,130 +1,0 @@
-<%- include('../layouts/header', { 
-  title: 'Your Containers - MIE',
-  breadcrumbs: [
-    { label: 'Sites', url: '/sites' },
-    { label: site.name, url: `/sites/${site.id}/containers` }
-  ],
-  colWidth: 'col-12 col-lg-10',
-  req
-}) %>
-
-<div class="card shadow-sm">
-  <div class="card-body">
-    <div class="d-flex justify-content-between align-items-center mb-3">
-      <h4 class="card-title mb-0">Your Containers - <%= site.name %></h4>
-      <a class="btn btn-primary" href="/sites/<%= site.id %>/containers/new" role="button">New Container</a>
-    </div>
-
-    <div class="table-responsive">
-      <table class="table table-hover table-bordered align-middle mb-0">
-        <thead class="table-light">
-          <tr>
-            <th>Hostname</th>
-            <th>Status</th>
-            <th>IPv4</th>
-            <th>Template</th>
-            <th>Node</th>
-            <th>SSH Port</th>
-            <th>HTTP</th>
-            <th>Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          <% if (rows && rows.length) { %>
-            <% rows.forEach(r => { %>
-              <tr>
-                <td><%= r.hostname %></td>
-                <td>
-                  <% if (r.status === 'running') { %>
-                    <span class="badge bg-success" aria-label="Container is running">Running</span>
-                  <% } else if (r.status === 'pending') { %>
-                    <span class="badge bg-warning text-dark" aria-label="Container creation pending">
-                      <span class="spinner-border spinner-border-sm me-1" role="status" aria-hidden="true"></span>
-                      Pending
-                    </span>
-                  <% } else if (r.status === 'creating') { %>
-                    <span class="badge bg-info text-dark" aria-label="Container is being created">
-                      <span class="spinner-border spinner-border-sm me-1" role="status" aria-hidden="true"></span>
-                      Creating
-                    </span>
-                  <% } else if (r.status === 'failed') { %>
-                    <span class="badge bg-danger" aria-label="Container creation failed">Failed</span>
-                  <% } else if (r.status === 'restarting') { %>
-                    <span class="badge bg-info text-dark" aria-label="Container is restarting">
-                      <span class="spinner-border spinner-border-sm me-1" role="status" aria-hidden="true"></span>
-                      Restarting
-                    </span>
-                  <% } else { %>
-                    <span class="badge bg-secondary" aria-label="Container status unknown"><%= r.status || 'Unknown' %></span>
-                  <% } %>
-                  <% if (r.creationJobId && (r.status === 'pending' || r.status === 'creating' || r.status === 'failed')) { %>
-                    <a href="/jobs/<%= r.creationJobId %>" class="btn btn-link btn-sm p-0 ms-1" aria-label="View job details">
-                      <i class="bi bi-info-circle"></i> Details
-                    </a>
-                  <% } %>
-                </td>
-                <td><%= r.ipv4Address || '-' %></td>
-                <td><%= r.template || '-' %></td>
-                <td>
-                  <% if (r.nodeApiUrl && r.containerId) { %>
-                    <a href="<%= r.nodeApiUrl %>#v1:0:=lxc%2F<%= r.containerId %>:4::::::=consolejs:" target="_blank" rel="noopener noreferrer" aria-label="Open <%= r.nodeName %> console for container <%= r.hostname %>"><%= r.nodeName %> <i class="bi bi-box-arrow-up-right"></i></a>
-                  <% } else { %>
-                    <%= r.nodeName %>
-                  <% } %>
-                </td>
-                <td>
-                  <% if (r.sshPort && r.sshHost) { %>
-                    <%= r.sshPort %>
-                    <a href="vscode://vscode-remote/ssh-remote+<%= req.session.user %>@<%= r.sshHost %>:<%= r.sshPort %>/" target="_blank" rel="noopener noreferrer" title="Open in VS Code" aria-label="Open SSH in VS Code for container <%= r.hostname %>" class="ms-2"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 256 256" aria-hidden="true" style="vertical-align: -0.125em;"><path fill="#0065A9" d="M246.94 27.64l-52.75-25.4c-6.1-2.94-13.4-1.7-18.2 3.1L3.33 162.77c-4.64 4.24-4.64 11.55.01 15.78l14.1 12.82c3.8 3.46 9.53 3.71 13.62.61L239 34.23c6.98-5.3 17-.32 17 8.44v-.61c0-6.15-3.52-11.75-9.06-14.42z"/><path fill="#007ACC" d="M246.94 228.36l-52.75 25.4c-6.1 2.94-13.4 1.7-18.2-3.09L3.33 93.23c-4.64-4.24-4.64-11.55.01-15.78l14.1-12.82c3.8-3.46 9.53-3.71 13.62-.61L239 221.77c6.98 5.3 17 .32 17-8.44v.61c0 6.15-3.52 11.75-9.06 14.42z"/><path fill="#1F9CF0" d="M194.2 253.76c-6.11 2.94-13.4 1.7-18.2-3.1 5.9 5.91 16 1.73 16-6.62V11.96c0-8.35-10.1-12.53-16-6.63 4.8-4.79 12.09-6.03 18.2-3.1l52.74 25.36c5.54 2.67 9.07 8.27 9.07 14.42v197.33c0 6.15-3.53 11.75-9.07 14.42z"/></svg></a>
-                    <a href="ssh://<%= req.session.user %>@<%= r.sshHost %>:<%= r.sshPort %>" target="_blank" rel="noopener noreferrer" title="Open SSH in terminal" aria-label="Open SSH terminal for container <%= r.hostname %>" class="ms-2"><i class="bi bi-terminal"></i></a>
-                  <% } else { %>
-                    <%= r.sshPort || '-' %>
-                  <% } %>
-                </td>
-                <td>
-                  <% if (r.httpEntries && r.httpEntries.length > 0) { %>
-                    <% r.httpEntries.forEach((entry, i) => { %>
-                      <% if (i > 0) { %><br><% } %>
-                      <% if (entry.externalUrl) { %>
-                        <a href="<%= entry.externalUrl %>" target="_blank" rel="noopener noreferrer" aria-label="Open external URL for container <%= r.hostname %>"><%= entry.port %> <i class="bi bi-box-arrow-up-right"></i></a>
-                      <% } else { %>
-                        <%= entry.port %>
-                      <% } %>
-                    <% }) %>
-                  <% } else { %>
-                    -
-                  <% } %>
-                </td>
-                <td>
-                  <% if (r.status === 'running' || r.status === 'failed') { %>
-                    <a href="/sites/<%= site.id %>/containers/<%= r.id %>/edit" class="btn btn-primary btn-sm" aria-label="Edit services for container <%= r.hostname %>">Edit</a>
-                    <form method="POST" action="/sites/<%= site.id %>/containers/<%= r.id %>" style="display: inline;">
-                      <input type="hidden" name="_method" value="PUT">
-                      <input type="hidden" name="restart" value="true">
-                      <button type="submit" class="btn btn-warning btn-sm" aria-label="Restart container <%= r.hostname %>">Restart</button>
-                    </form>
-                  <% } %>
-                  <form method="POST" action="/sites/<%= site.id %>/containers/<%= r.id %>" style="display: inline;" onsubmit="return confirm('Are you sure you want to delete container <%= r.hostname %>? This action cannot be undone.');">
-                    <input type="hidden" name="_method" value="DELETE">
-                    <button type="submit" class="btn btn-danger btn-sm" aria-label="Delete container <%= r.hostname %>">Delete</button>
-                  </form>
-                </td>
-              </tr>
-            <% }) %>
-          <% } else { %>
-            <tr>
-              <td colspan="8" class="text-muted text-center py-4">
-                No containers found. Click "New Container" to create your first one.
-              </td>
-            </tr>
-          <% } %>
-        </tbody>
-      </table>
-    </div>
-  </div>
-  <div class="card-footer text-muted text-center small">
-    &copy; 2025 Medical Informatics Engineering
-  </div>
-</div>
-
-<%- include('../layouts/footer') %>
