--- create-a-container/views/external-domains/form.ejs
@@ -1,174 +1,0 @@
-<%- include('../layouts/header', { 
-  title: (isEdit ? 'Edit External Domain' : 'New External Domain') + ' - MIE',
-  breadcrumbs: [
-    { label: 'External Domains', url: '/external-domains' },
-    { label: isEdit ? 'Edit' : 'New', url: '#' }
-  ],
-  colWidth: 'col-lg-8',
-  req
-}) %>
-
-<div class="card shadow-sm">
-  <div class="card-body">
-    <h4 class="card-title mb-4"><%= isEdit ? 'Edit External Domain' : 'Create New External Domain' %></h4>
-
-    <form method="POST" action="<%= isEdit ? `/external-domains/${externalDomain.id}` : '/external-domains' %>">
-      <% if (isEdit) { %>
-        <input type="hidden" name="_method" value="PUT">
-      <% } %>
-      
-      <div class="mb-3">
-        <label for="name" class="form-label">Domain Name <span class="text-danger">*</span></label>
-        <input 
-          type="text" 
-          class="form-control" 
-          id="name" 
-          name="name" 
-          value="<%= externalDomain ? externalDomain.name : '' %>"
-          required
-          placeholder="example.com"
-        >
-        <div class="form-text">The fully qualified domain name (e.g., example.com)</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="siteId" class="form-label">Default Site</label>
-        <select class="form-select" id="siteId" name="siteId" aria-label="Select default site">
-          <option value="">None</option>
-          <% sites.forEach(site => { %>
-            <option value="<%= site.id %>" <%= (externalDomain && externalDomain.siteId === site.id) ? 'selected' : '' %>><%= site.name %></option>
-          <% }) %>
-        </select>
-        <div class="form-text">The site whose DNS is assumed pre-configured (e.g., wildcard A record). Cross-site services on other sites will have Cloudflare A records created automatically.</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="acmeEmail" class="form-label">ACME Email</label>
-        <input 
-          type="email" 
-          class="form-control" 
-          id="acmeEmail" 
-          name="acmeEmail" 
-          value="<%= externalDomain ? externalDomain.acmeEmail : '' %>"
-          placeholder="admin@example.com"
-        >
-        <div class="form-text">Email address for ACME certificate registration and renewal notifications</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="acmeDirectoryUrl" class="form-label">ACME Directory</label>
-        <select 
-          class="form-select" 
-          id="acmeDirectoryUrl" 
-          name="acmeDirectoryUrl"
-        >
-          <option value="https://acme-v02.api.letsencrypt.org/directory" 
-            <%= (externalDomain && externalDomain.acmeDirectoryUrl === 'https://acme-v02.api.letsencrypt.org/directory') || !externalDomain ? 'selected' : '' %>>
-            Let's Encrypt (Production)
-          </option>
-          <option value="https://acme-staging-v02.api.letsencrypt.org/directory"
-            <%= externalDomain && externalDomain.acmeDirectoryUrl === 'https://acme-staging-v02.api.letsencrypt.org/directory' ? 'selected' : '' %>>
-            Let's Encrypt (Staging)
-          </option>
-          <option value="https://acme.zerossl.com/v2/DV90"
-            <%= externalDomain && externalDomain.acmeDirectoryUrl === 'https://acme.zerossl.com/v2/DV90' ? 'selected' : '' %>>
-            ZeroSSL
-          </option>
-        </select>
-        <div class="form-text">
-          Select the ACME certificate authority. ZeroSSL will automatically retrieve EAB credentials using your email.
-        </div>
-      </div>
-
-      <div class="mb-3">
-        <label for="cloudflareApiEmail" class="form-label">Cloudflare API Email</label>
-        <input 
-          type="email" 
-          class="form-control" 
-          id="cloudflareApiEmail" 
-          name="cloudflareApiEmail" 
-          value="<%= externalDomain ? externalDomain.cloudflareApiEmail : '' %>"
-          placeholder="user@example.com"
-        >
-        <div class="form-text">Cloudflare account email for DNS-01 challenge validation and cross-site DNS record management</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="cloudflareApiKey" class="form-label">Cloudflare API Key</label>
-        <input 
-          type="password" 
-          class="form-control" 
-          id="cloudflareApiKey" 
-          name="cloudflareApiKey" 
-          value=""
-          placeholder="<%= isEdit ? '(unchanged)' : '' %>"
-          autocomplete="off"
-        >
-        <div class="form-text">
-          API token with DNS:Edit permission. <%= isEdit ? 'Leave blank to keep existing API key. Enter new key to update.' : '' %>
-        </div>
-      </div>
-
-      <div class="mb-3">
-        <label for="authServer" class="form-label">Auth Server URL</label>
-        <input 
-          type="url" 
-          class="form-control" 
-          id="authServer" 
-          name="authServer" 
-          value="<%= externalDomain ? (externalDomain.authServer || '') : '' %>"
-          placeholder="https://manager.example.com"
-        >
-        <div class="form-text">
-          URL of an authentication server that supports the NGINX
-          <a href="https://nginx.org/en/docs/http/ngx_http_auth_request_module.html" target="_blank"><code>auth_request</code></a>
-          protocol. Must implement <code>GET /verify</code> (return 2xx if authenticated, 401 otherwise)
-          and <code>GET /login?redirect=&lt;url&gt;</code> for unauthenticated users.
-          The server must be on a subdomain of this external domain for session cookie sharing.
-        </div>
-        <div id="authServerWarning" class="form-text text-warning" style="display: none;">
-          ⚠️ The auth server URL does not appear to be on a subdomain of this domain.
-          Cookie sharing for <code>auth_request</code> may not work correctly.
-        </div>
-      </div>
-
-      <div class="d-flex justify-content-between">
-        <a href="/external-domains" class="btn btn-secondary" aria-label="Cancel and return to external domains list">Cancel</a>
-        <button type="submit" class="btn btn-primary">
-          <%= isEdit ? 'Update External Domain' : 'Create External Domain' %>
-        </button>
-      </div>
-    </form>
-  </div>
-  <div class="card-footer text-muted text-center small">
-    &copy; 2025 Medical Informatics Engineering
-  </div>
-</div>
-
-<script>
-(function() {
-  const domainInput = document.getElementById('name');
-  const authServerInput = document.getElementById('authServer');
-  const warning = document.getElementById('authServerWarning');
-
-  function checkDomainMatch() {
-    const domain = (domainInput.value || '').trim().toLowerCase();
-    const url = (authServerInput.value || '').trim();
-    if (!domain || !url) { warning.style.display = 'none'; return; }
-    try {
-      const parsed = new URL(url);
-      const hostname = parsed.hostname.toLowerCase();
-      const matches = hostname === domain || hostname.endsWith('.' + domain);
-      warning.style.display = matches ? 'none' : 'block';
-    } catch {
-      warning.style.display = 'none';
-    }
-  }
-
-  domainInput.addEventListener('input', checkDomainMatch);
-  authServerInput.addEventListener('input', checkDomainMatch);
-  checkDomainMatch();
-})();
-</script>
-
-<%- include('../layouts/footer') %>
