--- create-a-container/views/layouts/footer.ejs
@@ -1,31 +1,0 @@
-      </main>
-    </div>
-  </div>
-
-  <!-- Footer with version and issue reporting -->
-  <footer class="version-footer">
-    <div class="version-info">
-      <span class="version-label">Version:</span>
-      <a href="<%= versionInfo.url %>" target="_blank" rel="noopener noreferrer" class="version-link">
-        <%= versionInfo.display %>
-      </a>
-      <span class="version-date">(<%= versionInfo.date %>)</span>
-    </div>
-    <div class="issue-report">
-      <a href="https://github.com/mieweb/opensource-server/issues/new?template=bug_report.yml&url=<%- encodeURIComponent(req.originalUrl || '/') %>&username=<%- encodeURIComponent(req?.session?.user || 'Not logged in') %>&version=<%- encodeURIComponent(versionInfo.display) %>" 
-         target="_blank" 
-         rel="noopener noreferrer" 
-         class="report-issue-link"
-         aria-label="Report an issue on GitHub">
-        <i class="bi bi-bug"></i> Report Issue
-      </a>
-    </div>
-  </footer>
-
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/js/bootstrap.bundle.min.js" integrity="sha384-FKyoEForCGlyvwx9Hj09JcYn3nv7wiPVlz7YYwJrWVcXK/BmnVDxM+D2scQbITxI" crossorigin="anonymous"></script>
-  <script>
-    const popoverTriggerList = document.querySelectorAll('[data-bs-toggle="popover"]');
-    const popoverList = [...popoverTriggerList].map(popoverTriggerEl => new bootstrap.Popover(popoverTriggerEl));
-  </script>
-</body>
-</html>
