--- create-a-container/views/login.ejs
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>MIE Container Creation Login</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="/style.css" rel="stylesheet">
-</head>
-<body class="page-center">
-    <div class="login-container">
-        <img src="logo.png" alt="Company Logo" class="logo">
-        <h2>Container Creation Login</h2>
-        
-        <% if (successMessages && successMessages.length > 0) { %>
-            <% successMessages.forEach(function(message) { %>
-                <div class="alert alert-success">
-                    <%= message %>
-                </div>
-            <% }) %>
-        <% } %>
-        
-        <% if (errorMessages && errorMessages.length > 0) { %>
-            <% errorMessages.forEach(function(message) { %>
-                <div class="alert alert-error">
-                    <%- message %>
-                </div>
-            <% }) %>
-        <% } %>
-
-        <% if (typeof warningMessages !== 'undefined' && warningMessages && warningMessages.length > 0) { %>
-            <% warningMessages.forEach(function(message) { %>
-                <div class="alert alert-warning">
-                    <%= message %>
-                </div>
-            <% }) %>
-        <% } %>
-        
-        <form method="POST" action="/login" id="loginForm">
-            <input type="hidden" name="redirect" value="<%= redirect %>">
-            
-            <label for="username">Username</label>
-            <input type="text" name="username" id="username" required>
-
-            <label for="password">Password</label>
-            <input type="password" name="password" id="password" required>
-
-            <input type="submit" value="Login" id="loginButton">
-        </form>
-        
-        <!-- Loading banner for 2FA waiting -->
-        <div id="loadingBanner" class="alert alert-info" style="display: none;">
-            <strong>⏳ Waiting for authentication...</strong>
-            <p>Please check your device and respond to the push notification.</p>
-        </div>
-        
-        <hr>
-        <div class="signup-section">
-            <a href="/reset-password">Forgot your password?</a>
-        </div>
-        <hr>
-        <div class="signup-section">
-            <strong>Don't have an account?</strong>
-            <a href="/register">Register</a>
-        </div>
-        <div class="footer">
-            &copy; 2025 Medical Informatics Engineering
-        </div>
-    </div>
-    
-    <script>
-    // Wait for DOM to be fully loaded before attaching event listeners
-    document.addEventListener('DOMContentLoaded', function() {
-        // Show loading banner when form is submitted
-        document.getElementById('loginForm').addEventListener('submit', function(e) {
-            // Show loading banner
-            document.getElementById('loadingBanner').style.display = 'block';
-            // Disable submit button to prevent double submission
-            document.getElementById('loginButton').disabled = true;
-            document.getElementById('loginButton').value = 'Authenticating...';
-        });
-    });
-    </script>
-</body>
-</html>
