--- create-a-container/views/nodes/form.ejs
@@ -1,194 +1,0 @@
-<%- include('../layouts/header', { 
-  title: (isEdit ? 'Edit Node' : 'New Node') + ' - MIE',
-  breadcrumbs: [
-    { label: 'Sites', url: '/sites' },
-    { label: site.name, url: `/sites/${site.id}/nodes` },
-    { label: isEdit ? 'Edit' : 'New', url: '#' }
-  ],
-  colWidth: 'col-lg-8',
-  req
-}) %>
-
-<div class="card shadow-sm">
-  <div class="card-body">
-    <h4 class="card-title mb-4"><%= isEdit ? 'Edit Node' : 'Create New Node' %></h4>
-
-    <form method="POST" action="<%= isEdit ? `/sites/${site.id}/nodes/${node.id}` : `/sites/${site.id}/nodes` %>">
-      <% if (isEdit) { %>
-        <input type="hidden" name="_method" value="PUT">
-      <% } %>
-      <div class="mb-3">
-        <label for="name" class="form-label">Node Name <span class="text-danger">*</span></label>
-        <input 
-          type="text" 
-          class="form-control" 
-          id="name" 
-          name="name" 
-          value="<%= node ? node.name : '' %>" 
-          required
-          placeholder="e.g., intern-phxdc-pve1"
-        >
-        <div class="form-text">The Proxmox node name (must be unique)</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="ipv4Address" class="form-label">IPv4 Address</label>
-        <input 
-          type="text" 
-          class="form-control" 
-          id="ipv4Address" 
-          name="ipv4Address" 
-          value="<%= node && node.ipv4Address ? node.ipv4Address : '' %>"
-          placeholder="e.g., 10.15.0.4"
-          pattern="^((25[0-5]|(2[0-4]|1\d|[1-9]|)\d)\.?\b){4}$"
-        >
-        <div class="form-text">The node's IPv4 address (optional)</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="apiUrl" class="form-label">API URL</label>
-        <input 
-          type="url" 
-          class="form-control" 
-          id="apiUrl" 
-          name="apiUrl" 
-          value="<%= node && node.apiUrl ? node.apiUrl : '' %>"
-          placeholder="e.g., https://10.15.0.4:8006"
-        >
-        <div class="form-text">The Proxmox API endpoint URL (optional)</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="tokenId" class="form-label">Username</label>
-        <input 
-          type="text" 
-          class="form-control" 
-          id="tokenId" 
-          name="tokenId" 
-          value="<%= node && node.tokenId ? node.tokenId : '' %>"
-          placeholder="e.g., root@pam"
-          autocomplete="off"
-        >
-        <div class="form-text">Proxmox username (e.g., root@pam)</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="secret" class="form-label">Password</label>
-        <input 
-          type="password" 
-          class="form-control" 
-          id="secret" 
-          name="secret" 
-          value=""
-          placeholder="<%= isEdit ? '(unchanged)' : '' %>"
-          autocomplete="off"
-        >
-        <div class="form-text">
-          <%= isEdit ? 'Leave blank to keep existing password. Enter new password to update.' : 'Proxmox password' %>
-        </div>
-      </div>
-
-      <div class="mb-3">
-        <label for="tlsVerify" class="form-label">TLS Verification</label>
-        <select class="form-select" id="tlsVerify" name="tlsVerify">
-          <option value="" <%= (!node || node.tlsVerify === null) ? 'selected' : '' %>>Not Set (Default: Verify)</option>
-          <option value="true" <%= (node && node.tlsVerify === true) ? 'selected' : '' %>>Enable (Verify Certificates)</option>
-          <option value="false" <%= (node && node.tlsVerify === false) ? 'selected' : '' %>>Disable (Skip Verification)</option>
-        </select>
-        <div class="form-text">Whether to verify TLS certificates when connecting to this node</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="imageStorage" class="form-label">Template Storage</label>
-        <input 
-          type="text" 
-          class="form-control" 
-          id="imageStorage" 
-          name="imageStorage" 
-          value="<%= node && node.imageStorage ? node.imageStorage : 'local' %>"
-          placeholder="e.g., local"
-          list="imageStorageList"
-          autocomplete="off"
-        >
-        <datalist id="imageStorageList"></datalist>
-        <div class="form-text">Storage for CT Template images used when building containers</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="volumeStorage" class="form-label">Volume Storage</label>
-        <input 
-          type="text" 
-          class="form-control" 
-          id="volumeStorage" 
-          name="volumeStorage" 
-          value="<%= node && node.volumeStorage ? node.volumeStorage : 'local-lvm' %>"
-          placeholder="e.g., local-lvm"
-          list="volumeStorageList"
-          autocomplete="off"
-        >
-        <datalist id="volumeStorageList"></datalist>
-        <div class="form-text">Storage for container root filesystems</div>
-      </div>
-
-      <div class="mb-3">
-        <label for="networkBridge" class="form-label">Network Bridge</label>
-        <input 
-          type="text" 
-          class="form-control" 
-          id="networkBridge" 
-          name="networkBridge" 
-          value="<%= node && node.networkBridge ? node.networkBridge : 'vmbr0' %>"
-          placeholder="e.g., vmbr0"
-        >
-        <div class="form-text">Proxmox network bridge used when creating containers</div>
-      </div>
-
-      <div class="mb-3 form-check">
-        <input 
-          type="checkbox" 
-          class="form-check-input" 
-          id="nvidiaAvailable" 
-          name="nvidiaAvailable" 
-          value="true"
-          <%= (node && node.nvidiaAvailable) ? 'checked' : '' %>
-          aria-label="NVIDIA GPU available on this node"
-        >
-        <label class="form-check-label" for="nvidiaAvailable">NVIDIA Available</label>
-        <div class="form-text">Enable if this node has NVIDIA drivers and nvidia-container-toolkit installed</div>
-      </div>
-      </div>
-
-      <div class="d-flex justify-content-between">
-        <a href="/sites/<%= site.id %>/nodes" class="btn btn-secondary" aria-label="Cancel and return to nodes list">Cancel</a>
-        <button type="submit" class="btn btn-primary">
-          <%= isEdit ? 'Update Node' : 'Create Node' %>
-        </button>
-      </div>
-    </form>
-  </div>
-  <div class="card-footer text-muted text-center small">
-    &copy; 2025 Medical Informatics Engineering
-  </div>
-</div>
-
-<%- include('../layouts/footer') %>
-
-<% if (isEdit) { %>
-<script>
-(function() {
-  fetch('/sites/<%= site.id %>/nodes/<%= node.id %>/storages')
-    .then(res => res.json())
-    .then(storages => {
-      const imageDatalist = document.getElementById('imageStorageList');
-      const volumeDatalist = document.getElementById('volumeStorageList');
-      storages.forEach(s => {
-        const option = document.createElement('option');
-        option.value = s.name;
-        imageDatalist.appendChild(option);
-        volumeDatalist.appendChild(option.cloneNode(true));
-      });
-    })
-    .catch(() => {});
-})();
-</script>
-<% } %>
