--- create-a-container/views/nodes/index.ejs
@@ -1,84 +1,0 @@
-<%- include('../layouts/header', { 
-  title: 'Nodes - MIE',
-  breadcrumbs: [
-    { label: 'Sites', url: '/sites' },
-    { label: site.name, url: `/sites/${site.id}/nodes` }
-  ],
-  req
-}) %>
-
-<div class="card shadow-sm">
-  <div class="card-body">
-    <div class="d-flex justify-content-between align-items-center mb-3">
-      <h4 class="card-title mb-0">Proxmox Nodes for <%= site.name %></h4>
-      <div>
-        <a class="btn btn-secondary me-2" href="/sites/<%= site.id %>/nodes/import" role="button" aria-label="Import nodes from external API">Import Nodes</a>
-        <a class="btn btn-primary" href="/sites/<%= site.id %>/nodes/new" role="button" aria-label="Create new Proxmox node">New Node</a>
-      </div>
-    </div>
-
-    <div class="table-responsive">
-      <table class="table table-hover table-bordered align-middle mb-0">
-        <thead class="table-light">
-          <tr>
-            <th>Name</th>
-            <th>IPv4 Address</th>
-            <th>API URL</th>
-            <th>TLS Verify</th>
-            <th>Containers</th>
-            <th>Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          <% if (rows && rows.length) { %>
-            <% rows.forEach(r => { %>
-              <tr>
-                <td><%= r.name %></td>
-                <td><%= r.ipv4Address || '-' %></td>
-                <td><%= r.apiUrl || '-' %></td>
-                <td>
-                  <% if (r.tlsVerify === true) { %>
-                    <span class="badge bg-success">Yes</span>
-                  <% } else if (r.tlsVerify === false) { %>
-                    <span class="badge bg-warning text-dark">No</span>
-                  <% } else { %>
-                    <span class="badge bg-secondary">Not Set</span>
-                  <% } %>
-                </td>
-                <td><%= r.containerCount %></td>
-                <td>
-                  <a href="/sites/<%= site.id %>/nodes/<%= r.id %>/edit" class="btn btn-sm btn-outline-primary" aria-label="Edit node <%= r.name %>">Edit</a>
-                  <form method="POST" action="/sites/<%= site.id %>/nodes/<%= r.id %>" style="display: inline;" onsubmit="return confirm('Are you sure you want to delete node <%= r.name %>? This action cannot be undone.');">
-                    <input type="hidden" name="_method" value="DELETE">
-                    <button
-                      type="submit"
-                      class="btn btn-danger btn-sm"
-                      <% if (r.containerCount > 0) { %>
-                        disabled
-                        aria-disabled="true"
-                        aria-label="Cannot delete node with associated containers"
-                      <% } %>
-                    >
-                      Delete
-                    </button>
-                  </form>
-                </td>
-              </tr>
-            <% }) %>
-          <% } else { %>
-            <tr>
-              <td colspan="5" class="text-muted text-center py-4">
-                No nodes found. Click "New Node" to create your first one.
-              </td>
-            </tr>
-          <% } %>
-        </tbody>
-      </table>
-    </div>
-  </div>
-  <div class="card-footer text-muted text-center small">
-    &copy; 2025 Medical Informatics Engineering
-  </div>
-</div>
-
-<%- include('../layouts/footer') %>
