--- create-a-container/views/reset-password/request.ejs
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Reset Password - MIE</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="/style.css" rel="stylesheet">
-</head>
-<body class="page-center">
-    <div class="login-container">
-        <img src="/logo.png" alt="Company Logo" class="logo">
-        <h2>Reset Password</h2>
-        
-        <% if (successMessages && successMessages.length > 0) { %>
-            <% successMessages.forEach(function(message) { %>
-                <div style="color: green; margin-bottom: 1rem; padding: 0.5rem; border: 1px solid green; background-color: #d4edda; border-radius: 4px;">
-                    <%= message %>
-                </div>
-            <% }) %>
-        <% } %>
-        
-        <% if (errorMessages && errorMessages.length > 0) { %>
-            <% errorMessages.forEach(function(message) { %>
-                <div style="color: red; margin-bottom: 1rem; padding: 0.5rem; border: 1px solid red; background-color: #f8d7da; border-radius: 4px;">
-                    <%= message %>
-                </div>
-            <% }) %>
-        <% } %>
-        
-        <p style="margin-bottom: 1.5rem; color: #666;">
-            Enter your username or email address and we'll send you instructions to reset your password.
-        </p>
-        
-        <form method="POST" action="/reset-password">
-            <label for="usernameOrEmail">Username or Email</label>
-            <input type="text" name="usernameOrEmail" id="usernameOrEmail" required autofocus>
-
-            <input type="submit" value="Send Reset Instructions">
-        </form>
-        
-        <hr>
-        
-        <div class="signup-section">
-            <a href="/login">Back to Login</a>
-        </div>
-        
-        <div class="footer">
-            &copy; 2025 Medical Informatics Engineering
-        </div>
-    </div>
-</body>
-</html>
