--- create-a-container/views/reset-password/reset.ejs
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Set New Password - MIE</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="/style.css" rel="stylesheet">
-</head>
-<body class="page-center">
-    <div class="login-container">
-        <img src="/logo.png" alt="Company Logo" class="logo">
-        <h2>Set New Password</h2>
-        
-        <% if (successMessages && successMessages.length > 0) { %>
-            <% successMessages.forEach(function(message) { %>
-                <div style="color: green; margin-bottom: 1rem; padding: 0.5rem; border: 1px solid green; background-color: #d4edda; border-radius: 4px;">
-                    <%= message %>
-                </div>
-            <% }) %>
-        <% } %>
-        
-        <% if (errorMessages && errorMessages.length > 0) { %>
-            <% errorMessages.forEach(function(message) { %>
-                <div style="color: red; margin-bottom: 1rem; padding: 0.5rem; border: 1px solid red; background-color: #f8d7da; border-radius: 4px;">
-                    <%= message %>
-                </div>
-            <% }) %>
-        <% } %>
-        
-        <p style="margin-bottom: 1rem; color: #666;">
-            Username: <strong><%= username %></strong>
-        </p>
-        
-        <p style="margin-bottom: 1.5rem; color: #666; font-size: 0.9rem;">
-            Enter your new password below. Password must be at least 8 characters long.
-        </p>
-        
-        <form method="POST" action="/reset-password/<%= token %>">
-            <label for="password">New Password</label>
-            <input type="password" name="password" id="password" required minlength="8" autofocus>
-
-            <label for="confirmPassword">Confirm New Password</label>
-            <input type="password" name="confirmPassword" id="confirmPassword" required minlength="8">
-
-            <input type="submit" value="Reset Password">
-        </form>
-        
-        <hr>
-        
-        <div class="signup-section">
-            <a href="/login">Back to Login</a>
-        </div>
-        
-        <div class="footer">
-            &copy; 2025 Medical Informatics Engineering
-        </div>
-    </div>
-</body>
-</html>
