--- create-a-container/views/users/index.ejs
@@ -1,77 +1,0 @@
-<%- include('../layouts/header', { 
-  title: 'Users - MIE',
-  breadcrumbs: [
-    { label: 'Users', url: '/users' }
-  ],
-  req
-}) %>
-<div class="card shadow-sm">
-  <div class="card-body">
-    <div class="d-flex justify-content-between align-items-center mb-3">
-      <h4 class="card-title mb-0">Users</h4>
-      <div>
-        <a class="btn btn-success me-2" href="/users/invite" role="button" aria-label="Invite user via email">Invite User</a>
-        <a class="btn btn-primary" href="/users/new" role="button" aria-label="Create new user">New User</a>
-      </div>
-    </div>
-
-    <div class="table-responsive">
-      <table class="table table-hover table-bordered align-middle mb-0">
-        <thead class="table-light">
-          <tr>
-            <th>UID</th>
-            <th>Username</th>
-            <th>Full Name</th>
-            <th>Email</th>
-            <th>Status</th>
-            <th>Groups</th>
-            <th>Admin</th>
-            <th>Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          <% if (rows.length === 0) { %>
-            <tr>
-              <td colspan="8" class="text-center text-muted">No users found</td>
-            </tr>
-          <% } else { %>
-            <% rows.forEach(row => { %>
-              <tr>
-                <td><%= row.uidNumber %></td>
-                <td><%= row.uid %></td>
-                <td><%= row.cn %></td>
-                <td><%= row.mail %></td>
-                <td>
-                  <% if (row.status === 'active') { %>
-                    <span class="badge bg-success">Active</span>
-                  <% } else if (row.status === 'pending') { %>
-                    <span class="badge bg-warning text-dark">Pending</span>
-                  <% } else { %>
-                    <span class="badge bg-secondary"><%= row.status %></span>
-                  <% } %>
-                </td>
-                <td><%= row.groups %></td>
-                <td>
-                  <% if (row.isAdmin) { %>
-                    <span class="badge bg-danger">Yes</span>
-                  <% } else { %>
-                    <span class="badge bg-secondary">No</span>
-                  <% } %>
-                </td>
-                <td>
-                  <a href="/users/<%= row.uidNumber %>/edit" class="btn btn-sm btn-outline-primary" aria-label="Edit user <%= row.uid %>">Edit</a>
-                  <form method="POST" action="/users/<%= row.uidNumber %>" style="display: inline;" onsubmit="return confirm('Are you sure you want to delete user <%= row.uid %>?');">
-                    <input type="hidden" name="_method" value="DELETE">
-                    <button type="submit" class="btn btn-danger btn-sm" aria-label="Delete user <%= row.uid %>">Delete</button>
-                  </form>
-                </td>
-              </tr>
-            <% }) %>
-          <% } %>
-        </tbody>
-      </table>
-    </div>
-  </div>
-</div>
-
-<%- include('../layouts/footer') %>
