--- create-a-container/views/users/invite.ejs
@@ -1,46 +1,0 @@
-<%- include('../layouts/header', { 
-  title: 'Invite User - MIE',
-  breadcrumbs: [
-    { label: 'Users', url: '/users' },
-    { label: 'Invite', url: '#' }
-  ],
-  colWidth: 'col-lg-6',
-  req
-}) %>
-
-<div class="card shadow-sm">
-  <div class="card-body">
-    <h4 class="card-title mb-4">Invite User</h4>
-    <p class="text-muted mb-4">
-      Send an invitation email to a new user. They will receive a link to register their account, 
-      and their account will be automatically activated upon registration.
-    </p>
-
-    <form method="POST" action="/users/invite">
-      <div class="mb-3">
-        <label for="email" class="form-label">Email Address <span class="text-danger">*</span></label>
-        <input 
-          type="email" 
-          class="form-control" 
-          id="email" 
-          name="email" 
-          required
-          placeholder="e.g., newuser@example.com"
-          aria-describedby="emailHelp"
-        >
-        <div id="emailHelp" class="form-text">
-          The invitation link will be sent to this email address and will expire in 24 hours.
-        </div>
-      </div>
-
-      <div class="d-flex gap-2">
-        <button type="submit" class="btn btn-success" aria-label="Send invitation email">
-          <i class="bi bi-envelope me-1"></i> Send Invitation
-        </button>
-        <a href="/users" class="btn btn-secondary">Cancel</a>
-      </div>
-    </form>
-  </div>
-</div>
-
-<%- include('../layouts/footer') %>
